pyproject.toml
[project]
name = "ai-powered-insurance-policy-information-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.24",
    "langchain-cohere>=0.4.4",
    "langchain-community>=0.3.22",
    "langchain-core>=0.3.55",
    "langchain-pinecone>=0.2.5",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ruff>=0.11.6",
]
